Move TenantId to separate class
