+++ account-management/Domain/Tenants/Tenant.cs
@@ -1,20 +1,6 @@
 using PlatformPlatform.Foundation.DomainModeling.Entities;
-using PlatformPlatform.Foundation.DomainModeling.Identity;
 
 namespace PlatformPlatform.AccountManagement.Domain.Tenants;
-
-public sealed record TenantId(long Value) : StronglyTypedId<TenantId>(Value)
-{
-    public override string ToString()
-    {
-        return Value.ToString();
-    }
-
-    public static explicit operator TenantId(string value)
-    {
-        return new TenantId(Convert.ToInt64(value));
-    }
-}
 
 public sealed class Tenant : AggregateRoot<TenantId>
 {
